requirements.txt
fastapi==0.115.13
uvicorn==0.15.0
transformers==4.52.4
torch==2.7.1
streamlit==1.45.1
python-dotenv==1.0.0